While this may be controversial for something as simple as a TODO application, I've opted to move us to the development version of GNOME TODO.
The reasoning for shifting to 3.91.1 is primarily due to it not being viable to patch issues with the EDS (Evolution Data Server) support and NotificationWidget that would cause a negative number of notifications. This was most noticable in my local usage with Nextcloud Tasks, which are provided over CalDav. Additionally, issues such as [this one](https://gitlab.gnome.org/GNOME/gnome-todo/issues/179) related to tasks from other lists being shown have not been backported to the 3.28 branch. 3.91.1 also resolves a multitude of use-after-free bugs that were present in GNOME Todo that would result in crashes.